+++ hr_addon/hr_addon/api/utils.py
@@ -1,315 +1,9 @@
 from __future__ import unicode_literals
 import frappe
 from frappe import _
-from frappe.utils.data import date_diff, time_diff_in_hours
-from frappe.utils import get_datetime, getdate, today, comma_sep, flt
+from frappe.utils.data import date_diff
+from frappe.utils import getdate, today, comma_sep
 from frappe.core.doctype.role.role import get_info_based_on_role
-
-
-def get_employee_checkin(employee,atime):
-    ''' select DATE('date time');'''
-    employee = employee
-    atime = atime
-    checkin_list = frappe.db.sql(
-        """
-        SELECT  name,log_type,time,skip_auto_attendance,attendance FROM `tabEmployee Checkin` 
-        WHERE employee='%s' AND DATE(time)= DATE('%s') ORDER BY time ASC
-        """%(employee,atime), as_dict=1
-    )
-    return checkin_list or []
-
-def get_employee_default_work_hour(employee,adate):
-    ''' weekly working hour'''
-    employee = employee
-    adate = adate    
-    #validate current or active FY year WHERE --
-    # AND YEAR(valid_from) = CAST(%(year)s as INT) AND YEAR(valid_to) = CAST(%(year)s as INT)
-    # AND YEAR(w.valid_from) = CAST(('2022-01-01') as INT) AND YEAR(w.valid_to) = CAST(('2022-12-30') as INT);
-    target_work_hours= frappe.db.sql(
-        """ 
-    SELECT w.name,w.employee,w.valid_from,w.valid_to,d.day,d.hours,d.break_minutes  FROM `tabWeekly Working Hours` w  
-    LEFT JOIN `tabDaily Hours Detail` d ON w.name = d.parent 
-    WHERE w.employee='%s' AND d.day = DAYNAME('%s') and w.valid_from <= '%s' and w.valid_to >= '%s' and w.docstatus = 1
-    """%(employee,adate,adate,adate), as_dict=1
-    )
-
-    if not target_work_hours:
-        frappe.throw(_('Please create Weekly Working Hours for the selected Employee:{0} first for date : {1}.').format(employee,adate))
-
-    if len(target_work_hours) > 1:
-        target_work_hours= "<br> ".join([frappe.get_desk_link("Weekly Working Hours", w.name) for w in target_work_hours])
-        frappe.throw(_('There exist multiple Weekly Working Hours exist for the Date <b>{0}</b>: <br>{1} <br>').format(adate, target_work_hours))
-
-    return target_work_hours[0]
-
-
-@frappe.whitelist()
-def get_actual_employee_log(aemployee, adate):
-    '''total actual log'''
-    employee_checkins = get_employee_checkin(aemployee,adate)
-    employee_default_work_hour = get_employee_default_work_hour(aemployee,adate)
-    is_date_in_holiday_list = date_is_in_holiday_list(aemployee,adate)
-    fields=["name", "no_break_hours", "set_target_hours_to_zero_when_date_is_holiday"]
-    weekly_working_hours = frappe.db.get_list(doctype="Weekly Working Hours", filters={"employee": aemployee}, fields=fields)    
-    is_target_hours_zero_on_holiday = len(weekly_working_hours) > 0 and weekly_working_hours[0]["set_target_hours_to_zero_when_date_is_holiday"] == 1
-      
-    # check empty or none
-    if employee_checkins:
-        no_break_hours = True if len(weekly_working_hours) > 0 and weekly_working_hours[0]["no_break_hours"] == 1 else False
-        new_workday = get_workday(employee_checkins, employee_default_work_hour, no_break_hours, is_target_hours_zero_on_holiday, is_date_in_holiday_list)
-        return new_workday
-    else :
-        view_employee_attendance = get_employee_attendance(aemployee, adate)
-        
-        break_minutes = employee_default_work_hour.break_minutes
-        expected_break_hours = flt(break_minutes / 60)
-        
-        if is_target_hours_zero_on_holiday and is_date_in_holiday_list:
-            new_workday = {
-                "target_hours": 0,
-                "total_target_seconds": 0,
-                "break_minutes": employee_default_work_hour.break_minutes,
-                "actual_working_hours": 0,
-                "hours_worked": 0,
-                "nbreak": 0,
-                "attendance": view_employee_attendance[0].name if len(view_employee_attendance) > 0 else "",
-                "break_hours": 0,
-                "total_work_seconds": 0,
-                "total_break_seconds": 0,
-                "employee_checkins": [],
-                "first_checkin": "",
-                "last_checkout": "",
-                "expected_break_hours": 0,
-            }
-        else:
-            new_workday = {
-                "target_hours": employee_default_work_hour.hours,
-                "total_target_seconds": employee_default_work_hour.hours * 60 * 60,
-                "break_minutes": employee_default_work_hour.break_minutes,
-                "actual_working_hours": -employee_default_work_hour.hours,
-                "manual_workday": 1,
-                "hours_worked": 0,
-                "nbreak": 0,
-                "attendance": view_employee_attendance[0].name if len(view_employee_attendance) > 0 else "",
-                "break_hours": 0,
-                "employee_checkins": [],
-                "expected_break_hours": expected_break_hours,
-            }
-
-    return new_workday
-
-    
-
-
-def get_workday(employee_checkins, employee_default_work_hour, no_break_hours, is_target_hours_zero_on_holiday,is_date_in_holiday_list=False):
-    hr_addon_settings = frappe.get_doc("HR Addon Settings")
-    is_break_from_checkins_with_swapped_hours = hr_addon_settings.workday_break_calculation_mechanism == "Break Hours from Employee Checkins" and hr_addon_settings.swap_hours_worked_and_actual_working_hours
-    new_workday = {}
-
-    hours_worked = 0.0
-    total_duration = 0
-   
-    # not pair of IN/OUT either missing
-    if len(employee_checkins)% 2 != 0:
-        hours_worked = -36.0
-        employee_checkin_message = ""
-        for d in employee_checkins:
-            employee_checkin_message += "<li>CheckIn Type:{0} for {1}</li>".format(d.log_type, frappe.get_desk_link("Employee Checkin", d.name))
-
-        #frappe.msgprint("CheckIns must be in pair for the given date:<ul>{}</ul>".format(employee_checkin_message))
-
-    if (len(employee_checkins) % 2 == 0):
-        # seperate 'IN' from 'OUT'
-        clockin_list = [get_datetime(kin.time) for x,kin in enumerate(employee_checkins) if x % 2 == 0]
-        clockout_list = [get_datetime(kout.time) for x,kout in enumerate(employee_checkins) if x % 2 != 0]
-
-        # get total worked hours
-        for i in range(len(clockin_list)):
-            wh = time_diff_in_hours(clockout_list[i],clockin_list[i])
-            hours_worked += float(str(wh))
-
-        # Calculate difference between first check-in and last checkout
-        if clockin_list and clockout_list:
-            first_checkin = clockin_list[0]
-            last_checkout = clockout_list[-1]  # Last element of clockout_list
-            total_duration = time_diff_in_hours(last_checkout, first_checkin) 
-
-        if is_break_from_checkins_with_swapped_hours:
-            total_duration, hours_worked = hours_worked, total_duration
-
-    default_break_minutes = employee_default_work_hour.break_minutes
-    default_break_hours = flt(default_break_minutes / 60)
-    target_hours = employee_default_work_hour.hours
-
-    if len(employee_checkins) % 2 == 0:
-        break_from_checkins = 0.0
-        for i in range(len(clockout_list) - 1):
-            wh = time_diff_in_hours(clockin_list[i + 1], clockout_list[i])
-            break_from_checkins += float(wh)
-
-        if hr_addon_settings.workday_break_calculation_mechanism == "Break Hours from Employee Checkins":
-            break_hours = break_from_checkins
-
-        elif hr_addon_settings.workday_break_calculation_mechanism == "Break Hours from Weekly Working Hours":
-            break_hours = default_break_hours
-
-        elif hr_addon_settings.workday_break_calculation_mechanism == "Break Hours from Weekly Working Hours if Shorter breaks":
-            if break_from_checkins <= default_break_hours:
-                break_hours = default_break_hours
-        else:
-            break_hours = 0.0
-
-    else:
-        break_hours = flt(-360.0)
-    
-    total_target_seconds = target_hours * 60 * 60
-    total_work_seconds = flt(hours_worked * 60 * 60)
-    expected_break_hours = flt(default_break_minutes / 60)
-    total_break_seconds = flt(break_hours * 60 * 60)
-    hours_worked = flt(hours_worked)
-
-    if is_break_from_checkins_with_swapped_hours:
-        #swapping for gall
-        if hours_worked > 0:
-            actual_working_hours = hours_worked - break_hours
-        else:    
-            actual_working_hours = total_duration - expected_break_hours
-
-    else:
-        if total_duration > 0:
-            actual_working_hours = total_duration - break_hours
-        else:    
-            actual_working_hours = hours_worked - expected_break_hours
-    attendance = employee_checkins[0].attendance if len(employee_checkins) > 0 else ""
-
-    if no_break_hours and hours_worked < 6 and not is_break_from_checkins_with_swapped_hours: # TODO: set 6 as constant
-        default_break_minutes = 0
-        total_break_seconds = 0
-        #expected_break_hours = 0
-        actual_working_hours = hours_worked
-
-    if is_target_hours_zero_on_holiday and is_date_in_holiday_list:
-        target_hours = 0
-        total_target_seconds = 0
-
-    #if comp_off_doc:
-    #    hours_worked = 0
-    #    actual_working_hours = 0  
-    #    #frappe.msgprint(frappe.get_desk_link("Leave Application", comp_off_doc) )
-    #    frappe.msgprint("The selected employee: {} has a Leave Application with the leave type: 'Freizeitausgleich (Nicht buchen!)' on the given date :{}.".format(aemployee,adate))
-
-    # if target_hours == 0:
-    #     expected_break_hours = 0
-    #     total_break_seconds = 0
-
-    new_workday.update({
-        "target_hours": target_hours,
-        "total_target_seconds": total_target_seconds,
-        "break_minutes": default_break_minutes,
-        "hours_worked": hours_worked,
-        "expected_break_hours": expected_break_hours,
-        "actual_working_hours": actual_working_hours,
-        "total_work_seconds": total_work_seconds,
-        "nbreak": 0,
-        "attendance": attendance,        
-        "break_hours": break_hours,
-        "total_break_seconds": total_break_seconds,
-        "employee_checkins":employee_checkins,
-    })
-
-    return new_workday
-
-
-@frappe.whitelist()
-def get_actual_employee_log_for_bulk_process(aemployee, adate):
-    employee_checkins = get_employee_checkin(aemployee, adate)
-    employee_default_work_hour = get_employee_default_work_hour(aemployee, adate)
-    is_date_in_holiday_list = date_is_in_holiday_list(aemployee, adate)
-
-    # Initialize 'fields' before it's used
-    fields = ["name", "no_break_hours", "set_target_hours_to_zero_when_date_is_holiday"]
-    
-    # Fetch weekly working hours using the 'fields' variable
-    weekly_working_hours = frappe.db.get_list(doctype="Weekly Working Hours", filters={"employee": aemployee}, fields=fields)
-    is_target_hours_zero_on_holiday = len(weekly_working_hours) > 0 and weekly_working_hours[0]["set_target_hours_to_zero_when_date_is_holiday"] == 1
-
-    if employee_checkins:
-        # Determine if 'no_break_hours' should be set to True or False
-        no_break_hours = True if len(weekly_working_hours) > 0 and weekly_working_hours[0]["no_break_hours"] == 1 else False
-        new_workday = get_workday(employee_checkins, employee_default_work_hour, no_break_hours, is_target_hours_zero_on_holiday, is_date_in_holiday_list)
-    else:
-        view_employee_attendance = get_employee_attendance(aemployee, adate)
-        
-        break_minutes = employee_default_work_hour.break_minutes
-        expected_break_hours = flt(break_minutes / 60)
-        
-        if is_target_hours_zero_on_holiday and is_date_in_holiday_list:
-            new_workday = {
-                "target_hours": 0,
-                "total_target_seconds": 0,
-                "break_minutes": employee_default_work_hour.break_minutes,
-                "actual_working_hours": 0,
-                "hours_worked": 0,
-                "nbreak": 0,
-                "attendance": view_employee_attendance[0].name if len(view_employee_attendance) > 0 else "",
-                "break_hours": 0,
-                "total_work_seconds": 0,
-                "total_break_seconds": 0,
-                "employee_checkins": [],
-                "first_checkin": "",
-                "last_checkout": "",
-                "expected_break_hours": 0,
-            }
-        else:
-            new_workday = {
-                "target_hours": employee_default_work_hour.hours,
-                "total_target_seconds": employee_default_work_hour.hours * 60 * 60,
-                "break_minutes": employee_default_work_hour.break_minutes,
-                "actual_working_hours": -employee_default_work_hour.hours,
-                "manual_workday": 1,
-                "hours_worked": 0,
-                "nbreak": 0,
-                "attendance": view_employee_attendance[0].name if len(view_employee_attendance) > 0 else "",
-                "break_hours": 0,
-                "employee_checkins": [],
-                "expected_break_hours": expected_break_hours,
-            }
-
-    return new_workday
-
-
-
-def get_employee_attendance(employee,atime):
-    ''' select DATE('date time');'''
-    employee = employee
-    atime = atime
-    
-    attendance_list = frappe.db.sql(
-        """
-        SELECT  name,employee,status,attendance_date,shift FROM `tabAttendance` 
-        WHERE employee='%s' AND DATE(attendance_date)= DATE('%s') AND docstatus = 1 ORDER BY attendance_date ASC
-        """%(employee,atime), as_dict=1
-    )
-    return attendance_list
-
-
-@frappe.whitelist()
-def date_is_in_holiday_list(employee, date):
-	holiday_list = frappe.db.get_value("Employee", employee, "holiday_list")
-	if not holiday_list:
-		frappe.msgprint(_("Holiday list not set in {0}").format(employee))
-		return False
-
-	holidays = frappe.db.sql(
-        """
-            SELECT holiday_date FROM `tabHoliday`
-            WHERE parent=%s AND holiday_date=%s
-        """,(holiday_list, getdate(date))
-    )
-
-	return len(holidays) > 0
-
-
 
 # ----------------------------------------------------------------------
 # WORK ANNIVERSARY REMINDERS SEND TO EMPLOYEES LIST IN HR-ADDON-SETTINGS
